插入一个新的Pass-05，其余Pass编号依次往后递增
