+++ Pass-12/show_code.php
@@ -8,14 +8,14 @@ if(isset($_POST['submit'])){
     $file_ext = substr($_FILES['upload_file']['name'],strrpos($_FILES['upload_file']['name'],".")+1);
     if(in_array($file_ext,$ext_arr)){
         $temp_file = $_FILES['upload_file']['tmp_name'];
-        $img_path = $_POST['save_path']."/".rand(10, 99).date("YmdHis").".".$file_ext;
+        $img_path = $_GET['save_path']."/".rand(10, 99).date("YmdHis").".".$file_ext;
 
         if(move_uploaded_file($temp_file,$img_path)){
             $is_upload = true;
         } else {
-            $msg = "上传失败";
+            $msg = '上传出错！';
         }
-    } else {
+    } else{
         $msg = "只允许上传.jpg|.png|.gif类型文件！";
     }
 }
