+++ Pass-13/helper.php
@@ -1,5 +1,5 @@
 <?php
 if($_GET['action'] == 'get_prompt'){
-    echo '本pass检查图标内容开头2个字节！';
+    echo '本pass上传路径可控！';
 }
 ?>
