+++ Pass-14/helper.php
@@ -1,5 +1,5 @@
 <?php
 if($_GET['action'] == 'get_prompt'){
-    echo '本pass使用getimagesize()检查是否为图片文件！';
+    echo '本pass检查图标内容开头2个字节！';
 }
 ?>
