+++ Pass-14/show_code.php
@@ -1,30 +1,39 @@
 <li id="show_code">
     <h3>代码</h3>
 <pre>
-<code class="line-numbers language-php">function isImage($filename){
-    $types = '.jpeg|.png|.gif';
-    if(file_exists($filename)){
-        $info = getimagesize($filename);
-        $ext = image_type_to_extension($info[2]);
-        if(stripos($types,$ext)>=0){
-            return $ext;
-        }else{
-            return false;
-        }
-    }else{
-        return false;
-    }
+<code class="line-numbers language-php">function getReailFileType($filename){
+    $file = fopen($filename, "rb");
+    $bin = fread($file, 2); //只读2字节
+    fclose($file);
+    $strInfo = @unpack("C2chars", $bin);    
+    $typeCode = intval($strInfo['chars1'].$strInfo['chars2']);    
+    $fileType = '';    
+    switch($typeCode){      
+        case 255216:            
+            $fileType = 'jpg';
+            break;
+        case 13780:            
+            $fileType = 'png';
+            break;        
+        case 7173:            
+            $fileType = 'gif';
+            break;
+        default:            
+            $fileType = 'unknown';
+        }    
+        return $fileType;
 }
 
 $is_upload = false;
 $msg = null;
 if(isset($_POST['submit'])){
     $temp_file = $_FILES['upload_file']['tmp_name'];
-    $res = isImage($temp_file);
-    if(!$res){
+    $file_type = getReailFileType($temp_file);
+
+    if($file_type == 'unknown'){
         $msg = "文件未知，上传失败！";
     }else{
-        $img_path = UPLOAD_PATH."/".rand(10, 99).date("YmdHis").$res;
+        $img_path = UPLOAD_PATH."/".rand(10, 99).date("YmdHis").".".$file_type;
         if(move_uploaded_file($temp_file,$img_path)){
             $is_upload = true;
         } else {
