+++ Pass-15/helper.php
@@ -1,5 +1,5 @@
 <?php
 if($_GET['action'] == 'get_prompt'){
-    echo '本pass使用exif_imagetype()检查是否为图片文件！';
+    echo '本pass使用getimagesize()检查是否为图片文件！';
 }
 ?>
