+++ Pass-16/index.php
@@ -3,88 +3,39 @@ include '../config.php';
 include '../head.php';
 include '../menu.php';
 
+function isImage($filename){
+    //需要开启php_exif模块
+    $image_type = exif_imagetype($filename);
+    switch ($image_type) {
+        case IMAGETYPE_GIF:
+            return "gif";
+            break;
+        case IMAGETYPE_JPEG:
+            return "jpg";
+            break;
+        case IMAGETYPE_PNG:
+            return "png";
+            break;    
+        default:
+            return false;
+            break;
+    }
+}
+
 $is_upload = false;
 $msg = null;
-if (isset($_POST['submit'])){
-    // 获得上传文件的基本信息，文件名，类型，大小，临时文件路径
-    $filename = $_FILES['upload_file']['name'];
-    $filetype = $_FILES['upload_file']['type'];
-    $tmpname = $_FILES['upload_file']['tmp_name'];
-
-    $target_path=UPLOAD_PATH.'/'.basename($filename);
-
-    // 获得上传文件的扩展名
-    $fileext= substr(strrchr($filename,"."),1);
-
-    //判断文件后缀与类型，合法才进行上传操作
-    if(($fileext == "jpg") && ($filetype=="image/jpeg")){
-        if(move_uploaded_file($tmpname,$target_path)){
-            //使用上传的图片生成新的图片
-            $im = imagecreatefromjpeg($target_path);
-
-            if($im == false){
-                $msg = "该文件不是jpg格式的图片！";
-                @unlink($target_path);
-            }else{
-                //给新图片指定文件名
-                srand(time());
-                $newfilename = strval(rand()).".jpg";
-                //显示二次渲染后的图片（使用用户上传图片生成的新图片）
-                $img_path = UPLOAD_PATH.'/'.$newfilename;
-                imagejpeg($im,$img_path);
-                @unlink($target_path);
-                $is_upload = true;
-            }
-        } else {
-            $msg = "上传出错！";
-        }
-
-    }else if(($fileext == "png") && ($filetype=="image/png")){
-        if(move_uploaded_file($tmpname,$target_path)){
-            //使用上传的图片生成新的图片
-            $im = imagecreatefrompng($target_path);
-
-            if($im == false){
-                $msg = "该文件不是png格式的图片！";
-                @unlink($target_path);
-            }else{
-                 //给新图片指定文件名
-                srand(time());
-                $newfilename = strval(rand()).".png";
-                //显示二次渲染后的图片（使用用户上传图片生成的新图片）
-                $img_path = UPLOAD_PATH.'/'.$newfilename;
-                imagepng($im,$img_path);
-
-                @unlink($target_path);
-                $is_upload = true;               
-            }
-        } else {
-            $msg = "上传出错！";
-        }
-
-    }else if(($fileext == "gif") && ($filetype=="image/gif")){
-        if(move_uploaded_file($tmpname,$target_path)){
-            //使用上传的图片生成新的图片
-            $im = imagecreatefromgif($target_path);
-            if($im == false){
-                $msg = "该文件不是gif格式的图片！";
-                @unlink($target_path);
-            }else{
-                //给新图片指定文件名
-                srand(time());
-                $newfilename = strval(rand()).".gif";
-                //显示二次渲染后的图片（使用用户上传图片生成的新图片）
-                $img_path = UPLOAD_PATH.'/'.$newfilename;
-                imagegif($im,$img_path);
-
-                @unlink($target_path);
-                $is_upload = true;
-            }
-        } else {
-            $msg = "上传出错！";
-        }
+if(isset($_POST['submit'])){
+    $temp_file = $_FILES['upload_file']['tmp_name'];
+    $res = isImage($temp_file);
+    if(!$res){
+        $msg = "文件未知，上传失败！";
     }else{
-        $msg = "只允许上传后缀为.jpg|.png|.gif的图片文件！";
+        $img_path = UPLOAD_PATH."/".rand(10, 99).date("YmdHis").".".$res;
+        if(move_uploaded_file($temp_file,$img_path)){
+            $is_upload = true;
+        } else {
+            $msg = "上传出错！";
+        }
     }
 }
 ?>
