+++ Pass-17/index.php
@@ -5,25 +5,86 @@ include '../menu.php';
 
 $is_upload = false;
 $msg = null;
+if (isset($_POST['submit'])){
+    // 获得上传文件的基本信息，文件名，类型，大小，临时文件路径
+    $filename = $_FILES['upload_file']['name'];
+    $filetype = $_FILES['upload_file']['type'];
+    $tmpname = $_FILES['upload_file']['tmp_name'];
 
-if(isset($_POST['submit'])){
-    $ext_arr = array('jpg','png','gif');
-    $file_name = $_FILES['upload_file']['name'];
-    $temp_file = $_FILES['upload_file']['tmp_name'];
-    $file_ext = substr($file_name,strrpos($file_name,".")+1);
-    $upload_file = UPLOAD_PATH . '/' . $file_name;
+    $target_path=UPLOAD_PATH.'/'.basename($filename);
 
-    if(move_uploaded_file($temp_file, $upload_file)){
-        if(in_array($file_ext,$ext_arr)){
-             $img_path = UPLOAD_PATH . '/'. rand(10, 99).date("YmdHis").".".$file_ext;
-             rename($upload_file, $img_path);
-             $is_upload = true;
-        }else{
-            $msg = "只允许上传.jpg|.png|.gif类型文件！";
-            unlink($upload_file);
+    // 获得上传文件的扩展名
+    $fileext= substr(strrchr($filename,"."),1);
+
+    //判断文件后缀与类型，合法才进行上传操作
+    if(($fileext == "jpg") && ($filetype=="image/jpeg")){
+        if(move_uploaded_file($tmpname,$target_path)){
+            //使用上传的图片生成新的图片
+            $im = imagecreatefromjpeg($target_path);
+
+            if($im == false){
+                $msg = "该文件不是jpg格式的图片！";
+                @unlink($target_path);
+            }else{
+                //给新图片指定文件名
+                srand(time());
+                $newfilename = strval(rand()).".jpg";
+                //显示二次渲染后的图片（使用用户上传图片生成的新图片）
+                $img_path = UPLOAD_PATH.'/'.$newfilename;
+                imagejpeg($im,$img_path);
+                @unlink($target_path);
+                $is_upload = true;
+            }
+        } else {
+            $msg = "上传出错！";
+        }
+
+    }else if(($fileext == "png") && ($filetype=="image/png")){
+        if(move_uploaded_file($tmpname,$target_path)){
+            //使用上传的图片生成新的图片
+            $im = imagecreatefrompng($target_path);
+
+            if($im == false){
+                $msg = "该文件不是png格式的图片！";
+                @unlink($target_path);
+            }else{
+                 //给新图片指定文件名
+                srand(time());
+                $newfilename = strval(rand()).".png";
+                //显示二次渲染后的图片（使用用户上传图片生成的新图片）
+                $img_path = UPLOAD_PATH.'/'.$newfilename;
+                imagepng($im,$img_path);
+
+                @unlink($target_path);
+                $is_upload = true;               
+            }
+        } else {
+            $msg = "上传出错！";
+        }
+
+    }else if(($fileext == "gif") && ($filetype=="image/gif")){
+        if(move_uploaded_file($tmpname,$target_path)){
+            //使用上传的图片生成新的图片
+            $im = imagecreatefromgif($target_path);
+            if($im == false){
+                $msg = "该文件不是gif格式的图片！";
+                @unlink($target_path);
+            }else{
+                //给新图片指定文件名
+                srand(time());
+                $newfilename = strval(rand()).".gif";
+                //显示二次渲染后的图片（使用用户上传图片生成的新图片）
+                $img_path = UPLOAD_PATH.'/'.$newfilename;
+                imagegif($im,$img_path);
+
+                @unlink($target_path);
+                $is_upload = true;
+            }
+        } else {
+            $msg = "上传出错！";
         }
     }else{
-        $msg = '上传出错！';
+        $msg = "只允许上传后缀为.jpg|.png|.gif的图片文件！";
     }
 }
 ?>
@@ -32,7 +93,11 @@ if(isset($_POST['submit'])){
     <ol>
         <li>
             <h3>任务</h3>
-            <p>上传一个<code>webshell</code>到服务器。</p>
+            <p>上传<code>图片马</code>到服务器。</p>
+            <p>注意：</p>
+            <p>1.保证上传后的图片马中仍然包含完整的<code>一句话</code>或<code>webshell</code>代码。</p>
+            <p>2.使用<a href="<?php echo INC_VUL_PATH;?>" target="_bank">文件包含漏洞</a>能运行图片马中的恶意代码。</p>
+            <p>3.图片马要<code>.jpg</code>,<code>.png</code>,<code>.gif</code>三种后缀都上传成功才算过关！</p>
         </li>
         <li>
             <h3>上传区</h3>
