+++ Pass-19/helper.php
@@ -1,5 +1,5 @@
 <?php
 if($_GET['action'] == 'get_prompt'){
-    echo '本pass的取文件名通过$_POST来获取。';
+    echo '需要代码审计！';
 }
 ?>
