+++ Pass-19/show_code.php
@@ -1,31 +1,124 @@
 <li id="show_code">
-    <h3>代码</h3>
+    <h3>index.php代码</h3>
 <pre>
-<code class="line-numbers language-php">$is_upload = false;
+<code class="line-numbers language-php">//index.php
+$is_upload = false;
 $msg = null;
-if (isset($_POST['submit'])) {
-    if (file_exists(UPLOAD_PATH)) {
-        $deny_ext = array("php","php5","php4","php3","php2","html","htm","phtml","pht","jsp","jspa","jspx","jsw","jsv","jspf","jtml","asp","aspx","asa","asax","ascx","ashx","asmx","cer","swf","htaccess");
-
-        $file_name = $_POST['save_name'];
-        $file_ext = pathinfo($file_name,PATHINFO_EXTENSION);
-
-        if(!in_array($file_ext,$deny_ext)) {
-            $temp_file = $_FILES['upload_file']['tmp_name'];
-            $img_path = UPLOAD_PATH . '/' .$file_name;
-            if (move_uploaded_file($temp_file, $img_path)) { 
-                $is_upload = true;
-            }else{
-                $msg = '上传出错！';
-            }
-        }else{
-            $msg = '禁止保存为该类型文件！';
-        }
-
-    } else {
-        $msg = UPLOAD_PATH . '文件夹不存在,请手工创建！';
+if (isset($_POST['submit']))
+{
+    require_once("./myupload.php");
+    $imgFileName =time();
+    $u = new MyUpload($_FILES['upload_file']['name'], $_FILES['upload_file']['tmp_name'], $_FILES['upload_file']['size'],$imgFileName);
+    $status_code = $u->upload(UPLOAD_PATH);
+    switch ($status_code) {
+        case 1:
+            $is_upload = true;
+            $img_path = $u->cls_upload_dir . $u->cls_file_rename_to;
+            break;
+        case 2:
+            $msg = '文件已经被上传，但没有重命名。';
+            break; 
+        case -1:
+            $msg = '这个文件不能上传到服务器的临时文件存储目录。';
+            break; 
+        case -2:
+            $msg = '上传失败，上传目录不可写。';
+            break; 
+        case -3:
+            $msg = '上传失败，无法上传该类型文件。';
+            break; 
+        case -4:
+            $msg = '上传失败，上传的文件过大。';
+            break; 
+        case -5:
+            $msg = '上传失败，服务器已经存在相同名称文件。';
+            break; 
+        case -6:
+            $msg = '文件无法上传，文件不能复制到目标目录。';
+            break;      
+        default:
+            $msg = '未知错误！';
+            break;
     }
 }
+
+//myupload.php
+class MyUpload{
+......
+......
+...... 
+  var $cls_arr_ext_accepted = array(
+      ".doc", ".xls", ".txt", ".pdf", ".gif", ".jpg", ".zip", ".rar", ".7z",".ppt",
+      ".html", ".xml", ".tiff", ".jpeg", ".png" );
+
+......
+......
+......  
+  /** upload()
+   **
+   ** Method to upload the file.
+   ** This is the only method to call outside the class.
+   ** @para String name of directory we upload to
+   ** @returns void
+  **/
+  function upload( $dir ){
+    
+    $ret = $this->isUploadedFile();
+    
+    if( $ret != 1 ){
+      return $this->resultUpload( $ret );
+    }
+
+    $ret = $this->setDir( $dir );
+    if( $ret != 1 ){
+      return $this->resultUpload( $ret );
+    }
+
+    $ret = $this->checkExtension();
+    if( $ret != 1 ){
+      return $this->resultUpload( $ret );
+    }
+
+    $ret = $this->checkSize();
+    if( $ret != 1 ){
+      return $this->resultUpload( $ret );    
+    }
+    
+    // if flag to check if the file exists is set to 1
+    
+    if( $this->cls_file_exists == 1 ){
+      
+      $ret = $this->checkFileExists();
+      if( $ret != 1 ){
+        return $this->resultUpload( $ret );    
+      }
+    }
+
+    // if we are here, we are ready to move the file to destination
+
+    $ret = $this->move();
+    if( $ret != 1 ){
+      return $this->resultUpload( $ret );    
+    }
+
+    // check if we need to rename the file
+
+    if( $this->cls_rename_file == 1 ){
+      $ret = $this->renameFile();
+      if( $ret != 1 ){
+        return $this->resultUpload( $ret );    
+      }
+    }
+    
+    // if we are here, everything worked as planned :)
+
+    return $this->resultUpload( "SUCCESS" );
+  
+  }
+......
+......
+...... 
+};
 </code>
 </pre>
 </li>
