+++ Pass-21/helper.php
@@ -1,0 +1,5 @@
+<?php
+if($_GET['action'] == 'get_prompt'){
+    echo 'Pass-20来源于CTF，请审计代码！';
+}
+?>
